Fix bug for deciding the winner
The problem lied in the fact that gameWinner needed to be declared in the global scope for it to work.
Otherwise scope of the variable depends on where the function is declared and not where it is called. That is why the gameWinner var inside of the playGame function CANNOT be used!